djangoMergeDocs/djangoMergeDocs/requirements.txt
Django              == 3.2.23
django-cors-headers == 4.1.0
PyPDF2              == 3.0.1
Pillow              == 9.5.0
aiohttp             == 3.8.6